# docs/openapi-dify.yaml
openapi: 3.0.3
info:
  title: Outdoor SaaS API
  version: 1.0.0
  description: |
    户外广告管理系统 REST API
    用于 Dify AI 助手集成

    ## 鉴权说明
    本 API 支持 API Key 鉴权方式，用于 Dify 自定义工具等外部系统调用。

    ### 获取 API Key
    请联系系统管理员获取 API Key。

    ### 使用方式
    在请求头中添加：`X-API-Key: your-api-key`

    ### Dify 配置
    在 Dify 中创建自定义工具时：
    1. 鉴权方式选择 "API Key"
    2. 鉴权字段填写 `X-API-Key`
    3. API Key 值填写管理员提供的密钥
servers:
  - url: http://192.168.31.34:16000
    description: 本地开发环境（frp内网穿透）
paths:
...
  /api/community/{id}:
    get:
      tags:
        - 社区管理
      summary: 根据ID查询社区
      description: 获取指定ID的社区详细信息
      operationId: get_community
      parameters:
        - name: id
          in: path
          required: true
          description: 社区ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultCommunity'
        '404':
          description: 社区不存在

    delete:
      tags:
        - 社区管理
      summary: 删除社区
      description: 根据ID删除指定社区
      operationId: delete_community
      parameters:
        - name: id
          in: path
          required: true
          description: 社区ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
...
  /api/community/filter/page:
    post:
      tags:
        - 社区管理
      summary: 分页查询社区
      description: 根据条件分页查询社区列表，支持多种查询条件组合
      operationId: filter_communities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageResultCommunity'

  # ========== 道闸管理 ==========
...
  /api/barrier-gate/community/{communityId}:
    get:
      tags:
        - 道闸管理
      summary: 根据社区查询道闸
      description: 获取指定社区下的所有道闸
      operationId: get_barrier_gates_by_community
      parameters:
        - name: communityId
          in: path
          required: true
          description: 社区ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功
...
  /api/plan/{id}:
    get:
      tags:
        - 方案管理
      summary: 根据ID查询方案
      description: 获取指定ID的方案详细信息
      operationId: get_plan
      parameters:
        - name: id
          in: path
          required: true
          description: 方案ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功

    delete:
      tags:
        - 方案管理
      summary: 删除方案
      description: 根据ID删除指定方案
      operationId: delete_plan
      parameters:
        - name: id
          in: path
          required: true
          description: 方案ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
...
  /api/plan/filter/page:
    post:
      tags:
        - 方案管理
      summary: 分页查询方案
      description: 根据条件分页查询方案列表，支持多种查询条件组合
      operationId: filter_plans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageResultPlan'

  # ========== 方案关联管理 ==========
  /api/plan-community/filter/page:
    post:
      tags:
        - 方案关联管理
      summary: 分页查询方案社区关联
      description: 分页查询方案与社区的关联关系，支持多种查询条件组合
      operationId: filter_plan_communities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanCommunityQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageResultPlanCommunity'
  /api/plan-community:
    post:
      tags:
        - 方案关联管理
      summary: 创建方案社区关联
      description: 将社区添加到方案
      operationId: create_plan_community
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanCommunity'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
  /api/plan-community/batch:
    post:
      tags:
        - 方案关联管理
      summary: 批量创建方案社区关联
      description: 批量将多个社区添加到方案，支持统一设置投放日期和需求数量
      operationId: batch_create_plan_community
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanCommunityBatchCreateDTO'
      responses:
        '200':
          description: 批量创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
  /api/plan-community/{id}:
    delete:
      tags:
        - 方案关联管理
      summary: 删除方案社区关联
      description: 从方案中移除社区
      operationId: delete_plan_community
      parameters:
        - name: id
          in: path
          required: true
          description: 关联ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
  /api/plan-barrier/filter/page:
    post:
      tags:
        - 方案关联管理
      summary: 分页查询方案道闸关联
      description: 分页查询方案与道闸的关联关系，支持多种查询条件组合
      operationId: filter_plan_barriers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanBarrierQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageResultPlanBarrier'
  /api/plan-barrier:
    post:
      tags:
        - 方案关联管理
      summary: 创建方案道闸关联
      description: 将道闸添加到方案
      operationId: create_plan_barrier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanBarrier'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
  /api/plan-barrier/batch:
    post:
      tags:
        - 方案关联管理
      summary: 批量创建方案道闸关联
      description: 批量将多个道闸添加到方案，支持统一设置投放日期和方案社区关联
      operationId: batch_create_plan_barrier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanBarrierBatchCreateDTO'
      responses:
        '200':
          description: 批量创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
  /api/plan-barrier/{id}:
    delete:
      tags:
        - 方案关联管理
      summary: 删除方案道闸关联
      description: 从方案中移除道闸
      operationId: delete_plan_barrier
      parameters:
        - name: id
          in: path
          required: true
          description: 关联ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 请输入 JWT Token，格式：Bearer {token}
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key 鉴权，用于 Dify 等外部工具调用。请联系管理员获取 API Key。
  schemas:
...
    ResultInteger:
      type: object
      properties:
        code:
          type: integer
          description: 响应码，200表示成功
        message:
          type: string
          description: 响应消息
        data:
          type: integer
          description: 返回的整数数据（通常为ID或影响行数）
        timestamp:
          type: integer
          format: int64
          description: 时间戳
...
    ResultCommunity:
      type: object
      properties:
        code:
          type: integer
          description: 响应码，200表示成功
        message:
          type: string
          description: 响应消息
        data:
          $ref: '#/components/schemas/Community'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
...
    CommunityQueryParam:
      type: object
      description: 社区信息分页查询参数，支持精确匹配、模糊搜索、范围查询、IN查询等多种查询方式
      allOf:
        - $ref: '#/components/schemas/BaseQueryParam'
      properties:
        communityNo:
          type: string
          description: 社区编号（精确匹配）
        communityNoLike:
          type: string
          description: 社区编号模糊搜索，支持部分匹配
        buildingName:
          type: string
          description: 楼栋名称（精确匹配）
        buildingNameLike:
          type: string
          description: 楼栋名称模糊搜索，支持部分匹配
        buildingAddress:
          type: string
          description: 楼栋地址（精确匹配）
        buildingAddressLike:
          type: string
          description: 楼栋地址模糊搜索，支持部分匹配
        city:
          type: string
          description: 城市（精确匹配）
        cityList:
          type: array
          description: 城市列表（IN查询），可同时查询多个城市
          items:
            type: string
        coordLatMin:
          type: number
          format: decimal
          description: 纬度最小值，用于地理位置范围查询
        coordLatMax:
          type: number
          format: decimal
          description: 纬度最大值，用于地理位置范围查询
        coordLngMin:
          type: number
          format: decimal
          description: 经度最小值，用于地理位置范围查询
        coordLngMax:
          type: number
          format: decimal
          description: 经度最大值，用于地理位置范围查询

    # ========== 道闸 ==========
...
    PlanQueryParam:
      type: object
      description: 投放方案分页查询参数，支持精确匹配、模糊搜索、IN查询、范围查询等多种查询方式
      allOf:
        - $ref: '#/components/schemas/BaseQueryParam'
      properties:
        planNo:
          type: string
          description: 方案编号（精确匹配）
        planNoLike:
          type: string
          description: 方案编号模糊搜索，支持部分匹配
        planName:
          type: string
          description: 方案名称（精确匹配）
        planNameLike:
          type: string
          description: 方案名称模糊搜索，支持部分匹配
        customer:
          type: string
          description: 客户名称（精确匹配）
        customerLike:
          type: string
          description: 客户名称模糊搜索，支持部分匹配
        releaseStatus:
          type: integer
          description: 投放状态（精确匹配），具体状态码含义请参考枚举定义
        releaseStatusList:
          type: array
          description: 投放状态列表（IN查询），可同时查询多个投放状态
          items:
            type: integer
        salesType:
          type: integer
          description: 销售类型（精确匹配），具体类型码含义请参考枚举定义
        salesTypeList:
          type: array
          description: 销售类型列表（IN查询），可同时查询多个销售类型
          items:
            type: integer
        mediaType:
          type: integer
          description: 媒体类型（精确匹配），具体类型码含义请参考枚举定义
        mediaTypeList:
          type: array
          description: 媒体类型列表（IN查询），可同时查询多个媒体类型
          items:
            type: integer
        releaseDateBeginFrom:
          type: string
          format: date
          description: 投放开始日期范围起点（大于等于），用于筛选投放开始日期在此日期之后的数据
        releaseDateBeginTo:
          type: string
          format: date
          description: 投放开始日期范围终点（小于等于），用于筛选投放开始日期在此日期之前的数据
        releaseDateEndFrom:
          type: string
          format: date
          description: 投放结束日期范围起点（大于等于），用于筛选投放结束日期在此日期之后的数据
        releaseDateEndTo:
          type: string
          format: date
          description: 投放结束日期范围终点（小于等于），用于筛选投放结束日期在此日期之前的数据
        salesPerson:
          type: string
          description: 销售人员（精确匹配）
        salesPersonLike:
          type: string
          description: 销售人员模糊搜索，支持部分匹配
        budgetMin:
          type: integer
          format: int64
          description: 预算最小值（大于等于），用于筛选预算在此金额以上的方案
        budgetMax:
          type: integer
          format: int64
          description: 预算最大值（小于等于），用于筛选预算在此金额以下的方案
        estimatedReachMin:
          type: integer
          description: 预计触达人数最小值（大于等于），用于筛选预计触达人数在此数值以上的方案
        estimatedReachMax:
          type: integer
          description: 预计触达人数最大值（小于等于），用于筛选预计触达人数在此数值以下的方案

    # ========== 分页响应 ==========
...
    ResultPageResultCommunity:
      type: object
      description: 社区分页查询响应
      properties:
        code:
          type: integer
          description: 响应码，200表示成功
        message:
          type: string
          description: 响应消息
        data:
          type: object
          description: 分页结果
          properties:
            pageNum:
              type: integer
              description: 当前页码
            pageSize:
              type: integer
              description: 每页大小
            total:
              type: integer
              format: int64
              description: 总记录数
            pages:
              type: integer
              description: 总页数
            list:
              type: array
              description: 社区列表
              items:
                $ref: '#/components/schemas/Community'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
...
    ResultPageResultPlan:
      type: object
      description: 方案分页查询响应
      properties:
        code:
          type: integer
          description: 响应码，200表示成功
        message:
          type: string
          description: 响应消息
        data:
          type: object
          description: 分页结果
          properties:
            pageNum:
              type: integer
              description: 当前页码
            pageSize:
              type: integer
              description: 每页大小
            total:
              type: integer
              format: int64
              description: 总记录数
            pages:
              type: integer
              description: 总页数
            list:
              type: array
              description: 方案列表
              items:
                $ref: '#/components/schemas/Plan'
        timestamp:
          type: integer
          format: int64
          description: 时间戳

    # ========== 方案关联 ==========
    PlanCommunity:
      type: object
      description: 方案社区关联实体
      properties:
        id:
          type: integer
          description: 关联ID
        planId:
          type: integer
          description: 方案ID
        communityId:
          type: integer
          description: 社区ID
        releaseDateBegin:
          type: string
          format: date
          description: 在该社区的投放开始日期
        releaseDateEnd:
          type: string
          format: date
          description: 在该社区的投放结束日期
        barrierRequiredQty:
          type: integer
          description: 该社区需要的道闸数量
        frameRequiredQty:
          type: integer
          description: 该社区需要的框架数量
        releaseStatus:
          type: integer
          description: 发布状态
    PlanCommunityBatchCreateDTO:
      type: object
      description: 方案社区关联批量创建参数，用于一次性将多个社区添加到方案
      properties:
        planId:
          type: integer
          description: 方案ID（必填），指定要将社区添加到哪个方案
        communityIds:
          type: array
          description: 社区ID列表（必填），要添加到方案的社区ID列表，至少包含一个
          items:
            type: integer
          minItems: 1
        releaseDateBegin:
          type: string
          format: date
          description: 投放开始日期（可选），默认为空
        releaseDateEnd:
          type: string
          format: date
          description: 投放结束日期（可选），默认为空
        barrierRequiredQty:
          type: integer
          description: 道闸需求数量（可选），每个社区需要的道闸数量，默认为10
          default: 10
        frameRequiredQty:
          type: integer
          description: 框架需求数量（可选），每个社区需要的框架数量，默认为10
          default: 10
        releaseStatus:
          type: integer
          description: 发布状态（可选），默认为空
      required:
        - planId
        - communityIds
    PlanCommunityQueryParam:
      type: object
      description: 方案社区关联分页查询参数，支持精确匹配、IN查询、范围查询等多种查询方式
      allOf:
        - $ref: '#/components/schemas/BaseQueryParam'
      properties:
        planId:
          type: integer
          description: 方案ID（精确匹配）
        planIdList:
          type: array
          description: 方案ID列表（IN查询），可同时查询多个方案
          items:
            type: integer
        communityId:
          type: integer
          description: 社区ID（精确匹配）
        communityIdList:
          type: array
          description: 社区ID列表（IN查询），可同时查询多个社区
          items:
            type: integer
        releaseStatus:
          type: integer
          description: 发布状态（精确匹配），具体状态码含义请参考枚举定义
        releaseStatusList:
          type: array
          description: 发布状态列表（IN查询），可同时查询多个发布状态
          items:
            type: integer
        barrierRequiredQty:
          type: integer
          description: 道闸需求数量（精确匹配）
        frameRequiredQty:
          type: integer
          description: 框架需求数量（精确匹配）
        releaseDateBeginFrom:
          type: string
          format: date
          description: 发布开始日期范围起点（大于等于），用于筛选发布开始日期在此日期之后的数据
        releaseDateBeginTo:
          type: string
          format: date
          description: 发布开始日期范围终点（小于等于），用于筛选发布开始日期在此日期之前的数据
        releaseDateEndFrom:
          type: string
          format: date
          description: 发布结束日期范围起点（大于等于），用于筛选发布结束日期在此日期之后的数据
        releaseDateEndTo:
          type: string
          format: date
          description: 发布结束日期范围终点（小于等于），用于筛选发布结束日期在此日期之前的数据
    ResultPageResultPlanCommunity:
      type: object
      description: 方案社区关联分页查询响应
      properties:
        code:
          type: integer
          description: 响应码，200表示成功
        message:
          type: string
          description: 响应消息
        data:
          type: object
          description: 分页结果
          properties:
            pageNum:
              type: integer
              description: 当前页码
            pageSize:
              type: integer
              description: 每页大小
            total:
              type: integer
              format: int64
              description: 总记录数
            pages:
              type: integer
              description: 总页数
            list:
              type: array
              description: 方案社区关联列表
              items:
                $ref: '#/components/schemas/PlanCommunity'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
    PlanBarrier:
      type: object
      description: 方案道闸关联实体
      properties:
        id:
          type: integer
          description: 关联ID
        planId:
          type: integer
          description: 方案ID
        barrierGateId:
          type: integer
          description: 道闸ID
        planCommunityId:
          type: integer
          description: 方案社区关联ID，表示该道闸属于方案中的哪个社区
        releaseDateBegin:
          type: string
          format: date
          description: 在该道闸的投放开始日期
        releaseDateEnd:
          type: string
          format: date
          description: 在该道闸的投放结束日期
        releaseStatus:
          type: integer
          description: 发布状态
...
    PlanBarrierQueryParam:
      type: object
      description: 方案道闸关联分页查询参数，支持精确匹配、IN查询、范围查询等多种查询方式
      allOf:
        - $ref: '#/components/schemas/BaseQueryParam'
      properties:
        planId:
          type: integer
          description: 方案ID（精确匹配）
        planIdList:
          type: array
          description: 方案ID列表（IN查询），可同时查询多个方案
          items:
            type: integer
        barrierGateId:
          type: integer
          description: 道闸ID（精确匹配）
        barrierGateIdList:
          type: array
          description: 道闸ID列表（IN查询），可同时查询多个道闸
          items:
            type: integer
        planCommunityId:
          type: integer
          description: 方案社区关联ID（精确匹配），用于查询特定方案社区下的道闸
        planCommunityIdList:
          type: array
          description: 方案社区关联ID列表（IN查询），可同时查询多个方案社区
          items:
            type: integer
        releaseStatus:
          type: integer
          description: 发布状态（精确匹配），具体状态码含义请参考枚举定义
        releaseStatusList:
          type: array
          description: 发布状态列表（IN查询），可同时查询多个发布状态
          items:
            type: integer
        releaseDateBeginFrom:
          type: string
          format: date
          description: 发布开始日期范围起点（大于等于），用于筛选发布开始日期在此日期之后的数据
        releaseDateBeginTo:
          type: string
          format: date
          description: 发布开始日期范围终点（小于等于），用于筛选发布开始日期在此日期之前的数据
        releaseDateEndFrom:
          type: string
          format: date
          description: 发布结束日期范围起点（大于等于），用于筛选发布结束日期在此日期之后的数据
        releaseDateEndTo:
          type: string
          format: date
          description: 发布结束日期范围终点（小于等于），用于筛选发布结束日期在此日期之前的数据
    ResultPageResultPlanBarrier:
      type: object
      description: 方案道闸关联分页查询响应
      properties:
        code:
          type: integer
          description: 响应码，200表示成功
        message:
          type: string
          description: 响应消息
        data:
          type: object
          description: 分页结果
          properties:
            pageNum:
              type: integer
              description: 当前页码
            pageSize:
              type: integer
              description: 每页大小
            total:
              type: integer
              format: int64
              description: 总记录数
            pages:
              type: integer
              description: 总页数
            list:
              type: array
              description: 方案道闸关联列表
              items:
                $ref: '#/components/schemas/PlanBarrier'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
...
security:
  - ApiKeyAuth: []